+++ examples/jest/jest.setup.js
@@ -1,3 +1,0 @@
-const { JestReporter } = require("kelonio");
-
-JestReporter.initializeKelonio();