+++ src/plugin/jestReporter.ts
@@ -1,0 +1,3 @@
+import { JestReporter } from "..";
+
+export = JestReporter;