+++ src/plugin/jestReporterSetup.ts
@@ -1,0 +1,3 @@
+import { JestReporter } from "..";
+
+JestReporter.initializeKelonio();