+++ src/plugin/karmaReporter.ts
@@ -1,0 +1,5 @@
+import { KarmaReporter } from "..";
+
+module.exports = {
+    "reporter:kelonio": ["type", KarmaReporter],
+};