+++ src/plugin/karmaReporterSetup.ts
@@ -1,0 +1,3 @@
+import { KarmaReporter } from "..";
+
+KarmaReporter.initializeKelonio();