+++ src/plugin/mochaReporter.ts
@@ -1,0 +1,3 @@
+import { MochaReporter } from "..";
+
+export = MochaReporter;